TugasAkhir2/requirements.txt
mysql-connector-python==8.0.33
protobuf==3.20.3
tabulate==0.9.0
